float -> double

+++ SLAU/Class/Matrix.h
@@ -2,17 +2,17 @@
 
 class Matrix
 {
-	float **begin;
+	double **begin;
 	int n, m;
 	int permut;
-	float det;
+	double det;
 public:
 	Matrix();
 	Matrix(int, int);
 	Matrix(const Matrix&);
-	Matrix(float** temp, int, int);
+	Matrix(double** temp, int, int);
 	~Matrix();
-	float GetDet();
+	double GetDet();
 	void Create(int, int); // создание массива внутри класса
 	void Create(); // Можно вытащить массив из файла
 	void CreateDiagonal(int); // Можно вытащить массив из файла
@@ -23,9 +23,9 @@ public:
 	void ReadFile(char*);
 	Matrix& Transp(void);
 	Matrix& operator+(Matrix&);
-	Matrix& operator+(float);
+	Matrix& operator+(double);
 	Matrix& operator-(Matrix&);
-	Matrix& operator-(float);
+	Matrix& operator-(double);
 	Matrix& operator=(Matrix&);
 	Matrix& operator*(Matrix&);
 	bool operator==(Matrix&);
@@ -34,27 +34,27 @@ public:
 	Matrix& MethodGauss_bycolumn(Matrix&);
 	Matrix& InverseGauss();
 	Matrix& MethodHoleckogo(Matrix&);
-	Matrix& MethodYakobi(Matrix&, Matrix&,const float);
+	Matrix& MethodYakobi(Matrix&, Matrix&,const double);
 	Matrix& MethodKVKor(Matrix&);
-	Matrix& MethodGaussZeydel(Matrix&, Matrix&, const float);
+	Matrix& MethodGaussZeydel(Matrix&, Matrix&, const double);
 	Matrix& Integer(void);
 	Matrix& Okrugl(void);
-	Matrix& MethodYakobi(float, float, float);
-	Matrix& MethodSeidel(Matrix&, const float);
-	Matrix& MethodNewton(Matrix&, const float);
+	Matrix& MethodYakobi(double, double, double);
+	Matrix& MethodSeidel(Matrix&, const double);
+	Matrix& MethodNewton(Matrix&, const double);
 	Matrix& Nevyazka3();
 	Matrix& MatrixByVector(Matrix&, Matrix&);
-	float operator[](int);
-	float F1_S(float,float,float);
-	float F2_S(float,float,float);
-	float F3_S(float,float,float);
-	float F1(float,float,float);
-	float F2(float,float,float);
-	float F3(float,float,float);
-	float max(Matrix&,Matrix&);
-	float max(float,float);
-	float Norma(void);
-	float Norma2Vector(void);
-	void append(float);
+	double operator[](int);
+	double F1_S(double,double,double);
+	double F2_S(double,double,double);
+	double F3_S(double,double,double);
+	double F1(double,double,double);
+	double F2(double,double,double);
+	double F3(double,double,double);
+	double max(Matrix&,Matrix&);
+	double max(double,double);
+	double Norma(void);
+	double Norma2Vector(void);
+	void append(double);
 };
